Create Model seeds

+++ db/migrate/20201112054317_create_orders.rb
@@ -1,7 +1,7 @@
 class CreateOrders < ActiveRecord::Migration[6.0]
   def change
     create_table :orders do |t|
-      t.decimal :amount
+      t.float :amount
       t.datetime :completed_at
 
       t.timestamps
